requirements.txt
mediapipe==0.10.9
opencv-python
numpy
pandas
tqdm
torch
matplotlib
scikit-learn
seaborn
pillow
python-dotenv
google-cloud-texttospeech
